+++ Tests/AdvancedOperationTests/Mocks.swift
@@ -355,6 +355,10 @@ final internal class MockObserver: OperationObserving {
     didProduceCount += 1
   }
 
+  func operationDidFailConditionsEvaluations(operation: AdvancedOperation, withErrors errors: [Error]) {
+    //TODO
+  }
+
 }
 
 // MARK: - AdvancedOperationQueueDelegate